pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "praxis-copilot"
version = "0.1.0"
description = "Praxis Copilot - investment research automation"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.28",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "pydantic>=2.0",
    "litellm>=1.0",
    "yfinance>=0.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]
